abutton: handle early checked state changes
Resend the button check message if the button was checked
early.  If we get a setChecked function call before the
child control is created, the message will disappear into
the void.  So connect to the childCreated signal and resend
if the checked state has changed.

This requires that we actually send the aboutToCreate and
created signals in AWidget for custom classes (such as
buttons), because we won't be intercepting their WM_CREATE
messages.