ascrollbar: recalculate maximum value
We can use the real page size if we're careful about what maximum value
we pass to Windows.  Setting the page size to 1 gives a bad UX under
wine (the trackbox is paper thin), so try harder to get it right.  How
this works is described in the accompanying comment.