abutton: don't disconnect slots from radio group
They will be disconnected anyway, so remove it from our list of
associated radio buttons and call it a day.

It was causing the following crash under wine when ending the program:
    Unhandled exception: page fault on read access to 0x0000001a in
    32-bit code (0x0000001a).
This was likely caused by disconnecting the aboutToBeDeleteObject
signal while it was still firing, resulting in the premature
destruction of the calling lambda.

Ironic.  The code was wrong even though it was not crashing under
Windows 7 32bit.